+++ EnemizerLibrary/EnemizerLibrary.csproj
@@ -83,6 +83,7 @@
     <Compile Include="DungeonProperties.cs" />
     <Compile Include="DungeonsRoomsSprites.cs" />
     <Compile Include="DungeonsSpritesRandomization.cs" />
+    <Compile Include="EnemyRandomizer\DungeonEnemyRandomizer.cs" />
     <Compile Include="EnemyRandomizer\Room.cs" />
     <Compile Include="EnemyRandomizer\RoomConstants.cs" />
     <Compile Include="EnemyRandomizer\Sprite.cs" />
